@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq;
 using System.Threading.Tasks;
 using CQRSlite.Domain;
 using CQRSlite.Domain.Exception;
@@ -12,17 +13,19 @@ namespace CQRSlite.Tests.Domain
         private readonly TestAggregate _aggregate;
         private readonly Repository _rep;
         private readonly Session _session;
-
+        private readonly TestInMemoryEventStore _eventStore;
+        private readonly Guid _id;
 
         public When_saving_stale_data()
         {
-            var eventStore = new TestInMemoryEventStore();
-            _rep = new Repository(eventStore);
+            _eventStore = new TestInMemoryEventStore();
+            _rep = new Repository(_eventStore);
             _session = new Session(_rep);
 
-            _aggregate = new TestAggregate(Guid.NewGuid());
+            _id = Guid.NewGuid();
+            _aggregate = new TestAggregate(_id);
             _aggregate.DoSomething();
-            Task.Run(() => _rep.Save(_aggregate)).Wait();
+            Task.Run(async () => await _rep.Save(_aggregate)).Wait();
         }
 
         [Fact]
@@ -38,6 +41,16 @@ namespace CQRSlite.Tests.Domain
             _aggregate.DoSomething();
             await _rep.Save(_aggregate);
             await Assert.ThrowsAsync<ConcurrencyException>(async () => await _session.Commit());
+        }
+
+        [Fact]
+        public async Task Should_update_version_number()
+        {
+            var aggregate = new TestAggregate(_id);
+            aggregate.DoSomething();
+            await _rep.Save(aggregate);
+
+            Assert.Equal(_eventStore.Events.Last().Version, _eventStore.Events.Count);
         }
     }
 }
