@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Threading;
 using System.Threading.Tasks;
 using CQRSlite.Domain;
 using CQRSlite.Events;
@@ -8,24 +9,24 @@ using Xunit;
 
 namespace CQRSlite.Tests.Snapshotting
 {
-    public class When_saving_a_snapshotable_aggregate_for_each_change
+    public class When_saving_a_snapshotable_aggregate_with_custom_interval
     {
         private readonly TestInMemorySnapshotStore _snapshotStore;
 	    private readonly ISession _session;
 	    private readonly TestSnapshotAggregate _aggregate;
 
-        public When_saving_a_snapshotable_aggregate_for_each_change()
+        public When_saving_a_snapshotable_aggregate_with_custom_interval()
         {
             IEventStore eventStore = new TestInMemoryEventStore();
             _snapshotStore = new TestInMemorySnapshotStore();
-            var snapshotStrategy = new DefaultSnapshotStrategy();
+            var snapshotStrategy = new DefaultSnapshotStrategy(23);
             var repository = new SnapshotRepository(_snapshotStore, snapshotStrategy, new Repository(eventStore), eventStore);
             _session = new Session(repository);
             _aggregate = new TestSnapshotAggregate(Guid.NewGuid());
             
             Task.Run(async () =>
             {
-                for (var i = 0; i < 150; i++)
+                for (var i = 0; i < 50; i++)
                 {
                     await _session.Add(_aggregate);
                     _aggregate.DoSomething();
@@ -35,9 +36,9 @@ namespace CQRSlite.Tests.Snapshotting
         }
 
         [Fact]
-        public void Should_snapshot_100th_change()
+        public void Should_snapshot_46th_change()
         {
-            Assert.Equal(100, _snapshotStore.SavedVersion);
+            Assert.Equal(46, _snapshotStore.SavedVersion);
         }
 
         [Fact]
@@ -49,7 +50,7 @@ namespace CQRSlite.Tests.Snapshotting
         [Fact]
         public async Task Should_get_aggregate_back_correct()
         {
-            Assert.Equal(150, (await _session.Get<TestSnapshotAggregate>(_aggregate.Id)).Number);
+            Assert.Equal(50, (await _session.Get<TestSnapshotAggregate>(_aggregate.Id)).Number);
         }
     }
 }
