@@ -73,7 +73,8 @@ namespace CQRSlite.Caching
                 {
                     aggregate = (T) await _cache.Get(aggregateId).ConfigureAwait(false);
                     var events = await _eventStore.Get(aggregateId, aggregate.Version, cancellationToken).ConfigureAwait(false);
-                    if (events.Any() && events.First().Version != aggregate.Version + 1)
+                    var firstEvent = events.FirstOrDefault();
+                    if (firstEvent != null && firstEvent.Version != aggregate.Version + 1)
                     {
                         await _cache.Remove(aggregateId).ConfigureAwait(false);
                     }
