@@ -1,11 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 using CQRSlite.Domain;
-#if NET452
-using System.Runtime.Caching;
-#else
 using Microsoft.Extensions.Caching.Memory;
-#endif
 
 namespace CQRSlite.Caching
 {
@@ -14,87 +10,55 @@ namespace CQRSlite.Caching
     /// </summary>
     public class MemoryCache : ICache
     {
-#if NET452
-        private readonly System.Runtime.Caching.MemoryCache _cache;
-        private Func<CacheItemPolicy> _policyFactory;
-#else
-        private readonly MemoryCacheEntryOptions _cacheOptions;
+        private Func<MemoryCacheEntryOptions> _optionsFactory;
         private readonly IMemoryCache _cache;
-#endif
 
         public MemoryCache()
         {
 
-#if NET452
-            _cache = System.Runtime.Caching.MemoryCache.Default;
-            _policyFactory = () => new CacheItemPolicy {
-                SlidingExpiration = TimeSpan.FromMinutes(15)
-            };
-#else
-            _cacheOptions = new MemoryCacheEntryOptions
+            _optionsFactory = () => new MemoryCacheEntryOptions
             {
                 SlidingExpiration = TimeSpan.FromMinutes(15)
             };
             _cache = new Microsoft.Extensions.Caching.Memory.MemoryCache(new MemoryCacheOptions());
-#endif
-
         }
 
         public Task<bool> IsTracked(Guid id)
         {
-#if NET452
-            return Task.FromResult(_cache.Contains(id.ToString()));
-#else
             return Task.FromResult(_cache.TryGetValue(id, out var o) && o != null);
-#endif
         }
 
         public Task Set(Guid id, AggregateRoot aggregate)
         {
-#if NET452
-            _cache.Add(id.ToString(), aggregate, _policyFactory.Invoke());
-#else
-            _cache.Set(id, aggregate, _cacheOptions);
-#endif
+            _cache.Set(id, aggregate, _optionsFactory.Invoke());
             return Task.FromResult(0);
         }
 
         public Task<AggregateRoot> Get(Guid id)
         {
-#if NET452
-            return Task.FromResult((AggregateRoot)_cache.Get(id.ToString()));
-#else
-            return Task.FromResult((AggregateRoot) _cache.Get(id));
-#endif
+            return Task.FromResult((AggregateRoot)_cache.Get(id));
         }
 
         public Task Remove(Guid id)
         {
-#if NET452
-            _cache.Remove(id.ToString());
-#else
             _cache.Remove(id);
-#endif
             return Task.FromResult(0);
         }
 
         public void RegisterEvictionCallback(Action<Guid> action)
         {
-#if NET452
-            _policyFactory = () => new CacheItemPolicy
+            _optionsFactory = _optionsFactory = () =>
             {
-                SlidingExpiration = TimeSpan.FromMinutes(15),
-                RemovedCallback = x =>
+                var options = new MemoryCacheEntryOptions
                 {
-                    action.Invoke(Guid.Parse(x.CacheItem.Key));
-                }
+                    SlidingExpiration = TimeSpan.FromMinutes(15)
+                };
+                options.RegisterPostEvictionCallback((key, value, reason, state) =>
+                {
+                    action.Invoke((Guid)key);
+                });
+                return options;
             };
-#else
-            _cacheOptions.RegisterPostEvictionCallback((key, value, reason, state) =>
-            {
-                action.Invoke((Guid) key);
-            });
-#endif
         }
     }
 }
