@@ -5,7 +5,7 @@ namespace CQRSlite.Domain.Exception
     public class EventsOutOfOrderException : System.Exception
     {
         public EventsOutOfOrderException(Guid id)
-            : base($"Eventstore gave events for aggregate {id} out of order")
+            : base($"EventStore gave events for aggregate {id} out of order")
         { }
     }
 }
